🧹 Extract geometry tags to src/physics/tags.ts and fix MAIN_WIRE_TAG duplicate
