+++ src/store/antennaStore.ts
@@ -21,45 +21,7 @@ import type {
   NetworkLoad,
   AntennaType,
 } from '../physics/types';
-import {
-  SWEEP_F_MIN_MHZ,
-  SWEEP_F_MAX_MHZ,
-  TRANSFORMER_CHOKE_OHMS,
-  ATU_COMPONENT_Q,
-  DEFAULT_FEEDLINE_ID,
-  DEFAULT_FEEDLINE_LENGTH_M,
-  DEFAULT_ATU_MAIN_FEEDLINE_LENGTH_M,
-  DEFAULT_GROUND_ID,
-  DEFAULT_WIRE_RADIUS_M,
-  SLOPING_V_MIN_TIP_Z_M,
-  findFeedlinePreset,
-  feedlineLossDb,
-  findGroundPreset,
-  referenceLength,
-  halfWaveLength,
-  MAIN_WIRE_TAG,
-  LEFT_LEG_TAG,
-  RIGHT_LEG_TAG,
-  FEED_BRIDGE_TAG,
-  FEEDLINE_SHIELD_TAG,
-  FEED_BRIDGE_LENGTH_M,
-  DELTA_BASE_TAG,
-  SLOPING_V_LEFT_STUB_TAG,
-  SLOPING_V_RIGHT_STUB_TAG,
-  SLOPING_V_STUB_BOTTOM_Z_M,
-  TERMINATED_DELTA_LEFT_BASE_TAG,
-  TERMINATED_DELTA_RIGHT_BASE_TAG,
-  TERMINATED_DELTA_BRIDGE_TAG,
-  VERTICAL_WHIP_TAG,
-  DEFAULT_WHIP_LENGTH_M,
-  INVERTED_L_VERTICAL_TAG,
-  FOLDED_DIPOLE_OPPOSITE_TAG,
-  FOLDED_DIPOLE_CONNECTOR_TAG,
-  FOLDED_DIPOLE_TERM_BRIDGE_TAG,
-  FOLDED_DIPOLE_DEFAULT_APERTURE_M,
-  FOLDED_DIPOLE_MAX_APERTURE_M,
-  FOLDED_DIPOLE_FEED_R_OHMS,
-} from '../physics/constants';
+import { SWEEP_F_MIN_MHZ, SWEEP_F_MAX_MHZ, TRANSFORMER_CHOKE_OHMS, ATU_COMPONENT_Q, DEFAULT_FEEDLINE_ID, DEFAULT_FEEDLINE_LENGTH_M, DEFAULT_ATU_MAIN_FEEDLINE_LENGTH_M, DEFAULT_GROUND_ID, DEFAULT_WIRE_RADIUS_M, SLOPING_V_MIN_TIP_Z_M, findFeedlinePreset, feedlineLossDb, findGroundPreset, referenceLength, halfWaveLength, FEED_BRIDGE_LENGTH_M, SLOPING_V_STUB_BOTTOM_Z_M, DEFAULT_WHIP_LENGTH_M, FOLDED_DIPOLE_DEFAULT_APERTURE_M, FOLDED_DIPOLE_MAX_APERTURE_M, FOLDED_DIPOLE_FEED_R_OHMS } from '../physics/constants';
 
 // Re-export geometry tags for UI and tests.
 export {
@@ -76,11 +38,12 @@ export {
   INVERTED_L_VERTICAL_TAG,
   FOLDED_DIPOLE_OPPOSITE_TAG,
   FOLDED_DIPOLE_CONNECTOR_TAG,
-  FOLDED_DIPOLE_TERM_BRIDGE_TAG,
-};
+  FOLDED_DIPOLE_TERM_BRIDGE_TAG
+} from '../physics/tags';
 import type { Theme } from '../utils/themeColors';
 import type { UnitSystem } from '../physics/units';
 import type { AtuMatchConfig } from '../physics/impedance';
+import { MAIN_WIRE_TAG, LEFT_LEG_TAG, RIGHT_LEG_TAG, FEED_BRIDGE_TAG, FEEDLINE_SHIELD_TAG, SLOPING_V_LEFT_STUB_TAG, SLOPING_V_RIGHT_STUB_TAG, TERMINATED_DELTA_LEFT_BASE_TAG, TERMINATED_DELTA_RIGHT_BASE_TAG, TERMINATED_DELTA_BRIDGE_TAG, VERTICAL_WHIP_TAG, INVERTED_L_VERTICAL_TAG, FOLDED_DIPOLE_OPPOSITE_TAG, FOLDED_DIPOLE_TERM_BRIDGE_TAG } from '../physics/tags';
 import {
   buildInvertedVWires,
   buildSlopingVWires,
@@ -1098,7 +1061,6 @@ function computeFeedlineLayout(
     },
   };
 }
-
 
 function buildGroundParams(state: AntennaState): GroundParams {
   // The height<=0 short-circuit makes the model "free space" when the user
