+++ tests/antennaStore.test.ts
@@ -1,33 +1,8 @@
 import { afterEach, describe, expect, it } from "vitest";
-import {
-  useAntennaStore,
-  buildWires,
-  selectSimulationInput,
-  selectAtuConfig,
-  selectSwrWindow,
-  legMultipleFromLength,
-  computeOptimalVAngleDeg,
-  recommendedTerminatingResistor,
-  SLOPING_V_DEFAULT_TERMINATION_OHMS,
-  TERMINATED_DELTA_DEFAULT_TERMINATION_OHMS,
-  LEFT_LEG_TAG,
-  RIGHT_LEG_TAG,
-  DELTA_BASE_TAG,
-  FEEDLINE_SHIELD_TAG,
-  VERTICAL_WHIP_TAG,
-  type AntennaState,
-} from "../src/store/antennaStore";
-import {
-  SWEEP_F_MIN_MHZ,
-  SWEEP_F_MAX_MHZ,
-  ATU_COMPONENT_Q,
-  findFeedlinePreset,
-  DEFAULT_WHIP_LENGTH_M,
-  VERTICAL_WHIP_BASE_GAP_M,
-  VERTICAL_WHIP_RADIAL_TAG,
-  VERTICAL_WHIP_RADIAL_COUNT,
-  FOLDED_DIPOLE_FEED_R_OHMS,
-} from "../src/physics/constants";
+import { useAntennaStore, buildWires, selectSimulationInput, selectAtuConfig, selectSwrWindow, legMultipleFromLength, computeOptimalVAngleDeg, recommendedTerminatingResistor, SLOPING_V_DEFAULT_TERMINATION_OHMS, TERMINATED_DELTA_DEFAULT_TERMINATION_OHMS, type AntennaState } from '../src/store/antennaStore';
+import { SWEEP_F_MIN_MHZ, SWEEP_F_MAX_MHZ, ATU_COMPONENT_Q, findFeedlinePreset, DEFAULT_WHIP_LENGTH_M, VERTICAL_WHIP_BASE_GAP_M, VERTICAL_WHIP_RADIAL_COUNT, FOLDED_DIPOLE_FEED_R_OHMS } from '../src/physics/constants';
+import { MAIN_WIRE_TAG, LEFT_LEG_TAG, RIGHT_LEG_TAG, DELTA_BASE_TAG, FEEDLINE_SHIELD_TAG, VERTICAL_WHIP_TAG } from '../src/physics/tags';
+import { VERTICAL_WHIP_RADIAL_TAG } from '../src/physics/tags';
 
 describe("antennaStore selectors", () => {
   describe("buildWires", () => {
@@ -288,7 +263,7 @@ describe("antennaStore selectors", () => {
       expect(input.frequencyMHz).toBe(14.1);
       expect(input.wires).toHaveLength(1);
       expect(input.wires[0].segments).toBe(11);
-      expect(input.excitation.wireTag).toBe(1);
+      expect(input.excitation.wireTag).toBe(MAIN_WIRE_TAG);
       expect(input.excitation.segment).toBe(6); // Math.ceil(11 / 2)
       expect(input.patternResolution.thetaSteps).toBe(37);
       expect(input.patternResolution.phiSteps).toBe(72);
@@ -499,7 +474,6 @@ describe("antennaStore selectors", () => {
     });
   });
 });
-
 
 describe("folded-dipole defaults", () => {
   // These switch the shared store's antenna type; put it back so the
